#128710

Related work items: #126534, #128633, #128637, #128710, #128755, #128762

+++ Logitude/Unifreight.Data/AmitalModel/Repsitories/SyncRecordRepository.cs
@@ -120,7 +120,7 @@ namespace Unifreight.Data.AmitalModel.Repsitories
             return groupRecord;
         }
 
-        public int GetFileNo(int tenant, int customsFileNo) =>
+        public int GetFileNo(int tenant, long customsFileNo) =>
             context.CCUFILEMs.Where(file => file.TENANT == tenant && file.CUSTOMFILENO == customsFileNo)
                 .Select(file => file.FILENO)
                 .FirstOrDefault();
